+++ weblate/trans/models/translation.py
@@ -21,12 +21,12 @@ import codecs
 import os
 import tempfile
 from datetime import datetime
-from typing import BinaryIO, List, Optional, Union
+from typing import BinaryIO, Dict, List, Optional, Union
 
 from django.core.cache import cache
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from django.db import IntegrityError, models, transaction
-from django.db.models import Q
+from django.db.models import F, Q
 from django.urls import reverse
 from django.utils import timezone
 from django.utils.functional import cached_property
@@ -51,6 +51,7 @@ from weblate.trans.models.unit import (
     STATE_TRANSLATED,
     Unit,
 )
+from weblate.trans.models.variant import Variant
 from weblate.trans.signals import store_post_load, vcs_pre_commit
 from weblate.trans.util import join_plural, split_plural
 from weblate.trans.validators import validate_check_flags
@@ -65,14 +66,10 @@ from weblate.utils.stats import GhostStats, TranslationStats
 class TranslationManager(models.Manager):
     def check_sync(self, component, lang, code, path, force=False, request=None):
         """Parse translation meta info and updates translation object."""
-        translation = self.get_or_create(
+        translation = component.translation_set.get_or_create(
             language=lang,
-            component=component,
             defaults={"filename": path, "language_code": code, "plural": lang.plural},
         )[0]
-        # Share component instance to improve performance
-        # and to properly process updated data.
-        translation.component = component
         if translation.filename != path or translation.language_code != code:
             force = True
             translation.filename = path
@@ -308,7 +305,14 @@ class Translation(
             report_error(cause="Translation parse error")
             self.component.handle_parse_error(exc, self)
 
-    def sync_unit(self, dbunits, updated, id_hash, unit, pos):
+    def sync_unit(
+        self,
+        dbunits: Dict[int, Unit],
+        updated: Dict[int, Unit],
+        id_hash: int,
+        unit,
+        pos: int,
+    ):
         try:
             newunit = dbunits[id_hash]
             is_new = False
@@ -316,6 +320,10 @@ class Translation(
             newunit = Unit(translation=self, id_hash=id_hash, state=-1)
             # Avoid fetching empty list of checks from the database
             newunit.all_checks = []
+            # Avoid fetching empty list of variants
+            newunit._prefetched_objects_cache = {
+                "defined_variants": Variant.objects.none()
+            }
             is_new = True
 
         newunit.update_from_unit(unit, pos, is_new)
@@ -326,9 +334,9 @@ class Translation(
         # - newly fuzzy
         # - source string changed
         if newunit.state < STATE_TRANSLATED and (
-            newunit.state != newunit.old_unit.state
+            newunit.state != newunit.old_unit["state"]
             or is_new
-            or newunit.source != newunit.old_unit.source
+            or newunit.source != newunit.old_unit["source"]
         ):
             self.was_new += 1
 
@@ -503,31 +511,33 @@ class Translation(
         return User.objects.get(pk=self.stats.last_author).get_author_name(email)
 
     @transaction.atomic
-    def commit_pending(self, reason, user, skip_push=False, force=False, signals=True):
+    def commit_pending(
+        self, reason: str, user, skip_push: bool = False, signals: bool = True
+    ):
         """Commit any pending changes."""
-        # Commit template first
-        if (
-            not self.is_source
-            and self.component.has_template()
-            and self.component.source_translation.needs_commit()
-        ):
-            self.component.source_translation.commit_pending(
-                reason, user, skip_push=skip_push, force=force, signals=signals
-            )
-
-        if not force and not self.needs_commit():
-            return False
-
-        self.log_info("committing pending changes (%s)", reason)
-
-        try:
-            store = self.store
-        except FileParseError as error:
-            report_error(cause="Failed to parse file on commit")
-            self.log_error("skipping commit due to error: %s", error)
-            return False
-
         with self.component.repository.lock:
+            # Commit template first
+            if (
+                not self.is_source
+                and self.component.has_template()
+                and self.component.source_translation.needs_commit()
+            ):
+                self.component.source_translation.commit_pending(
+                    reason, user, skip_push=skip_push, signals=signals
+                )
+
+            if not self.needs_commit():
+                return False
+
+            self.log_info("committing pending changes (%s)", reason)
+
+            try:
+                store = self.store
+            except FileParseError as error:
+                report_error(cause="Failed to parse file on commit")
+                self.log_error("skipping commit due to error: %s", error)
+                return False
+
             units = (
                 self.unit_set.filter(pending=True)
                 .prefetch_recent_content_changes()
@@ -582,9 +592,6 @@ class Translation(
         if self.component.file_format_cls.simple_filename:
             return [self.get_filename()]
         return self.store.get_filenames()
-
-    def repo_needs_commit(self):
-        return self.component.repository.needs_commit(self.filenames)
 
     def git_commit(
         self,
@@ -659,7 +666,7 @@ class Translation(
                 updated = True
 
                 # Optionally add unit to translation file.
-                # This has be done prior setting tatget as some formats
+                # This has be done prior setting target as some formats
                 # generate content based on target language.
                 if add:
                     store.add_unit(pounit.unit)
@@ -1025,7 +1032,8 @@ class Translation(
             )
             accepted += 1
         self.invalidate_cache()
-        self.component.update_variants()
+        if self.component.needs_variants_update:
+            self.component.update_variants()
         self.component.schedule_sync_terminology()
         self.component.update_source_checks()
         self.component.run_batched_checks()
@@ -1270,10 +1278,11 @@ class Translation(
                     state=STATE_TRANSLATED if bool(current_target) else STATE_EMPTY,
                     source_unit=source_unit,
                     id_hash=id_hash,
-                    position=0,
+                    position=translation.stats.all + 1,
                     **kwargs,
                 )
                 unit.is_batch_update = is_batch_update
+                unit.trigger_update_variants = False
                 try:
                     with transaction.atomic():
                         unit.save(force_insert=True)
@@ -1296,7 +1305,8 @@ class Translation(
                 result = unit
 
         if not is_batch_update:
-            component.update_variants()
+            if self.component.needs_variants_update:
+                component.update_variants()
             component.schedule_sync_terminology()
         return result
 
@@ -1311,15 +1321,29 @@ class Translation(
             previous_revision = self.component.repository.last_revision
             for translation in self.get_store_change_translations():
                 try:
+                    translation_unit = translation.unit_set.get(id_hash=unit.id_hash)
+                except ObjectDoesNotExist:
+                    continue
+                try:
                     pounit, add = translation.store.find_unit(unit.context, unit.source)
                 except UnitNotFound:
-                    return
+                    continue
                 if add:
-                    return
+                    continue
                 extra_files = translation.store.remove_unit(pounit.unit)
                 translation.addon_commit_files.extend(extra_files)
                 translation.drop_store_cache()
                 translation.git_commit(user, user.get_author_name(), store_hash=False)
+                # Adjust position as it will happen in most formats
+                translation.unit_set.filter(
+                    position__gt=translation_unit.position
+                ).update(position=F("position") - 1)
+            if self.is_source and not component.has_template():
+                # Adjust position is source language
+                self.unit_set.filter(position__gt=unit.position).update(
+                    position=F("position") - 1
+                )
+
             self.handle_store_change(request, user, previous_revision)
 
     @transaction.atomic
@@ -1352,8 +1376,6 @@ class Translation(
         extra = {}
         if isinstance(source, str):
             source = [source]
-        if isinstance(target, str):
-            target = [target]
         if not self.component.has_template():
             extra["source"] = join_plural(source)
         if not auto_context and self.unit_set.filter(context=context, **extra).exists():
@@ -1375,56 +1397,6 @@ class Translation(
                 explanation=explanation,
             )
             return
-        # Always load a new copy of store
-        store = self.load_store()
-        old_units = len(store.content_units)
-        # Add new unit
-        store.new_unit(context, source, target, skip_build=True)
-        # Serialize the content
-        handle = BytesIOMode("", b"")
-        # Catch serialization error
-        try:
-            store.save_content(handle)
-        except Exception as error:
-            raise ValidationError(_("Failed adding string: %s") % error)
-        handle.seek(0)
-        # Parse new file (check that it is valid)
-        try:
-            newstore = self.load_store(handle)
-        except Exception as error:
-            raise ValidationError(_("Failed adding string: %s") % error)
-        # Verify there is a single unit added
-        if len(newstore.content_units) != old_units + 1:
-            raise ValidationError(
-                _("Failed adding string: %s") % _("Failed to parse new string")
-            )
-        # Find newly added unit (it can be on any position), but we assume
-        # the storage has consistent ordering
-        unit = None
-        for pos, current in enumerate(newstore.content_units):
-            if pos >= old_units or (
-                current.source != store.content_units[pos].source
-                and current.context != store.content_units[pos].context
-            ):
-                unit = current
-                break
-        # Verify unit matches data
-        if unit is None:
-            raise ValidationError(
-                _("Failed adding string: %s") % _("Failed to parse new string")
-            )
-        created_source = split_plural(unit.source)
-        if unit.context != context and (
-            self.component.has_template()
-            or self.component.file_format_cls.set_context_bilingual
-        ):
-            raise ValidationError(
-                {"context": _('Context would be created as "%s"') % unit.context}
-            )
-        if created_source != source:
-            raise ValidationError(
-                {"source": _("Source would be created as %s") % created_source}
-            )
 
 
 class GhostTranslation: