+++ weblate/trans/tests/test_selenium.py
@@ -488,9 +488,14 @@ class SeleniumTests(BaseLiveServerTestCase, RegistrationTestMixin, TempDirMixin)
                 self.click("Dashboard")
 
         def wait_search():
+            time.sleep(0.1)
             WebDriverWait(self.driver, 15).until(
                 presence_of_element_located(
-                    (By.XPATH, '//div[@id="search-results"]//tr')
+                    (
+                        By.XPATH,
+                        '//div[@id="search-results"]'
+                        '//tbody[@class="unit-listing-body"]//tr',
+                    )
                 )
             )
 
@@ -747,11 +752,18 @@ class SeleniumTests(BaseLiveServerTestCase, RegistrationTestMixin, TempDirMixin)
         with self.wait_for_page_load():
             self.click(htmlid="engage-project")
 
-        # Add-ons
         self.click("Components")
         with self.wait_for_page_load():
             self.click("Language names")
+
+        # Repository
         self.click("Manage")
+        self.click("Repository maintenance")
+        time.sleep(0.2)
+        self.click("Manage")
+        self.screenshot("component-repository.png")
+
+        # Add-ons
         with self.wait_for_page_load():
             self.click("Add-ons")
         self.screenshot("addons.png")